tmf: Add keyboard shortcuts for Next/Previous Marker
Support '.' for Next Marker and ',' for Previous Marker.

Support Shift key for extending the selection, both with the keyboard
shortcut and with the action button.

Update the status line when Next/Previous Marker changes the selection.

Clarify the marker event source API to indicate that the previous and
next markers should be included in the marker list.

Change-Id: I1f435e9620030809f10f3788eca40521c8a745a9
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>