analysis: Fix IndexOutOfBoundsException in ResourcesView
This can happen if a new IRQ or SoftIRQ attribute is created
concurrently with a zoom thread that has already queried the state
system for a full state that does not include the new attribute.

The zoom thread can then provide the full state to the ResourcesView to
get the event list of an IRQ or SoftIRQ entry that did not exist at the
time that the full state was created.

Change-Id: I1c5f8a8332f44b44f1954b01c3503a844dedc22e
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>